レビュー内容修正

+++ app/controllers/items_controller.rb
@@ -38,9 +38,13 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    unless current_user.id == @item.user_id 
+      redirect_to root_path
+    else
+      item = Item.find(params[:id])
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
